check for illegal arguments in API
Check for a backslash, a double quote or semicolon. Why? one it's an
addiational way to verify that the  query is valid, secondly it's a
naive way to stop SQL injections, though it shouldn't matter since our
users don't have any write abilities and there's nothing but data to
query